requirements-local.txt
# Dipendenze per pre-processing locale
# pip install -r requirements-local.txt

pypdf2==3.0.1
pillow==10.1.0
pytesseract==0.3.10
chromadb==0.4.22
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2

# NOTA: Devi anche installare Tesseract OCR:
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr tesseract-ocr-ita